Safari-611.1.10/LayoutTests/fast/dom/HTMLSelectElement/listbox-items-client-rects-expected.txt

PASS listBoundingBoxes.length is 1
PASS listBoundingBoxes[0].left is listBoundingBox.left
PASS listBoundingBoxes[0].right is listBoundingBox.right
PASS listBoundingBoxes[0].top is listBoundingBox.top
PASS listBoundingBoxes[0].bottom is listBoundingBox.bottom
PASS listBoundingBoxes[0].width is listBoundingBox.width
PASS listBoundingBoxes[0].height is listBoundingBox.height
PASS groupBoundingBoxes.length is 1
PASS groupBoundingBoxes[0].left is groupBoundingBox.left
PASS groupBoundingBoxes[0].right is groupBoundingBox.right
PASS groupBoundingBoxes[0].top is groupBoundingBox.top
PASS groupBoundingBoxes[0].bottom is groupBoundingBox.bottom
PASS groupBoundingBoxes[0].width is groupBoundingBox.width
PASS groupBoundingBoxes[0].height is groupBoundingBox.height
PASS option1BoundingBoxes.length is 1
PASS option1BoundingBoxes[0].left is option1BoundingBox.left
PASS option1BoundingBoxes[0].right is option1BoundingBox.right
PASS option1BoundingBoxes[0].top is option1BoundingBox.top
PASS option1BoundingBoxes[0].bottom is option1BoundingBox.bottom
PASS option1BoundingBoxes[0].width is option1BoundingBox.width
PASS option1BoundingBoxes[0].height is option1BoundingBox.height
PASS option2BoundingBoxes.length is 1
PASS option2BoundingBoxes[0].left is option2BoundingBox.left
PASS option2BoundingBoxes[0].right is option2BoundingBox.right
PASS option2BoundingBoxes[0].top is option2BoundingBox.top
PASS option2BoundingBoxes[0].bottom is option2BoundingBox.bottom
PASS option2BoundingBoxes[0].width is option2BoundingBox.width
PASS option2BoundingBoxes[0].height is option2BoundingBox.height
PASS option3BoundingBoxes.length is 1
PASS option3BoundingBoxes[0].left is option3BoundingBox.left
PASS option3BoundingBoxes[0].right is option3BoundingBox.right
PASS option3BoundingBoxes[0].top is option3BoundingBox.top
PASS option3BoundingBoxes[0].bottom is option3BoundingBox.bottom
PASS option3BoundingBoxes[0].width is option3BoundingBox.width
PASS option3BoundingBoxes[0].height is option3BoundingBox.height
PASS groupBoundingBox.left is listBoundingBox.left + border
PASS groupBoundingBox.top is listBoundingBox.top + border
PASS groupBoundingBox.right is listBoundingBox.right - border
PASS groupBoundingBox.bottom is groupBoundingBox.top + 3 * optionHeight
PASS groupBoundingBox.width is listBoundingBox.width - 2 * border
PASS groupBoundingBox.height is 3 * optionHeight
PASS option1BoundingBox.left is listBoundingBox.left + border
PASS option1BoundingBox.top is listBoundingBox.top + border + optionHeight
PASS option1BoundingBox.right is listBoundingBox.right - border
PASS option1BoundingBox.bottom is option1BoundingBox.top + option1BoundingBox.height
PASS option1BoundingBox.width is listBoundingBox.width - 2 * border
PASS option2BoundingBox.left is listBoundingBox.left + border
PASS option2BoundingBox.top is listBoundingBox.top + border + 2 * optionHeight
PASS option2BoundingBox.right is listBoundingBox.right - border
PASS option2BoundingBox.bottom is option2BoundingBox.top + option2BoundingBox.height
PASS option2BoundingBox.width is listBoundingBox.width - 2 * border
PASS option3BoundingBox.left is listBoundingBox.left + border
PASS option3BoundingBox.top is listBoundingBox.top + border + 3 * optionHeight
PASS option3BoundingBox.right is listBoundingBox.right - border
PASS option3BoundingBox.bottom is option3BoundingBox.top + option3BoundingBox.height
PASS option3BoundingBox.width is listBoundingBox.width - 2 * border
PASS successfullyParsed is true

TEST COMPLETE
